Moved Xcode project search to its own struct for testing. Wrote test.
