+++ Parasol/main.swift
@@ -8,21 +8,7 @@
 
 import Foundation
 
-let fileManager = NSFileManager.defaultManager()
-
-do {
-    let files = try fileManager.contentsOfDirectoryAtPath(fileManager.currentDirectoryPath)
-    var xcodeProjectName: String?
-    print(files)
-    for file in files {
-        print(file)
-        print((file as NSString).pathExtension)
-        if (file as NSString).pathExtension == "xcodeproj" {
-            xcodeProjectName = file
-        }
-    }
-    print(xcodeProjectName)
-} catch {
+if let xcodeProjectName = XcodeProject.xcodeProjectName() {
     
 }
